fix deadlock when returning web3
upon calling DebugContract::new() the process would hang indefinitely.
After debugging, it was found web3 holds a mutex lock on the web3
struct, and testing or web3 library spawns multiple threads. Returning
was not possible, because accessing the mutex lock is not possible. This
means, web3 cannot be a member of the struct `DebugContract` and needs
to be re-instantiated wherever it is used.