Fix hang in ethdbg (A note on Synchronous calls using ganache-core testrpc)
since getting rid of using two processes (one for forked process testRPC,
the other for debugging) making a synchronous call to the TestRPC
will hang execution indefinitely. In this case, it was the call to
`this.web3.eth.coinbase` which hung execution. In these situations,
which will mostly happen in contract deployment, a forked worker process
will communicate with the TestRPC. It is important these worker
processes are made as efficient and succinct as possible.